+++ R/register_all_transpilers.R
@@ -34,6 +34,7 @@ register_all_transpilers <- function() {
 
   ## Domain-specific packages
   transpilers_for_package("futurize::add-on", package = "caret",          append_transpilers_for_caret)
+  transpilers_for_package("futurize::add-on", package = "DiceKriging",    append_transpilers_for_DiceKriging)
   transpilers_for_package("futurize::add-on", package = "fwb",            append_transpilers_for_fwb)
   transpilers_for_package("futurize::add-on", package = "gamlss",         append_transpilers_for_gamlss)
   transpilers_for_package("futurize::add-on", package = "glmnet",         append_transpilers_for_glmnet)